% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catapultR-package.R
\docType{package}
\name{catapultR-package}
\alias{catapultR}
\alias{catapultR-package}
\title{catapultR: Access to the Catapult OpenField Data}
\description{
Access to the Catapult data stored in the OpenField cloud and supplementary routines.
}
\keyword{internal}
