% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus.R
\name{eventsToFocus}
\alias{eventsToFocus}
\alias{activityToFocus}
\alias{periodToFocus}
\title{convert IMA events into Focus format}
\usage{
eventsToFocus(dfEvents, event_name, player_name, col = "blue", digits = 3)

activityToFocus(dfActivity, player_name = "", col = "blue")

periodToFocus(dfPeriods, player_name = "", col = "blue")
}
\arguments{
\item{dfEvents}{as returned by \code{\link{ofCloudGetActivityEvents}} or \code{\link{ofCloudGetPeriodEvents}}}

\item{event_name}{type of events in \code{dfEvents}, presumably from the set returned by \code{\link{ima_events}}}

\item{player_name}{name of the athlete associated with events}

\item{col}{color on the Focus timeline}

\item{digits}{the number of decimal places to round event attributes to}

\item{dfActivity}{an activity as returned from \code{\link{ofCloudGetActivities}}}

\item{dfPeriods}{periods as returned from \code{\link{ofCloudGetPeriods}}}
}
\value{
\code{eventsToFocus} returns a Focus-friendly data with the specified events.

\code{activityToFocus} returns a Focus-friendly data with the specified activity.

\code{periodToFocus} returns a Focus-friendly data with the specified periods.
}
\description{
Converts specified IMA events into a Focus-friendly data frame ready for \code{jsonlite::write_json}.
}
\section{Functions}{
\itemize{
\item \code{activityToFocus()}: converts activity description into a Focus-friendly data frame ready for \code{jsonlite::write_json}.

\item \code{periodToFocus()}: converts periods description into a Focus-friendly data frame ready for \code{jsonlite::write_json}.

}}
\examples{
\dontrun{
 dfActivities <- ofCloudGetActivities(token)
 dfEvents <- ofCloudGetActivityEvents(token, athlete_id, dfActivities$id[1], "ice_hockey_goaltender_movement")
 dfEvents <- dfEvents$ice_hockey_goaltender_movement
 dfEventsJSON <- eventsToFocus(dfEvents, "ice_hockey_goaltender_movement", "player1")
 jsonlite::write_json(dfEventsJSON, "iceHockeyFocus.JSON", pretty=TRUE)
 dfActivityJSON <- activityToFocus(dfActivities[1,])
 jsonlite::write_json(dfActivityJSON, "activity.JSON", pretty=TRUE)
 dfPeriods <- ofCloudGetPeriods(token, dfActivities$id[1])
 dfPeriodJSON <- periodToFocus(dfPeriods)
 jsonlite::write_json(dfPeriodJSON, "periods.JSON", pretty=TRUE)
 all.JSON <- dplyr::bind_rows(dfActivityJSON, dfPeriodJSON, dfEventsJSON) 
 all.JSON <- all.JSON[order(all.JSON$"tag-start"),]
 jsonlite::write_json(all.JSON, "all.JSON", pretty=TRUE)
}
}
