% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsIMA.R
\name{ima_events}
\alias{ima_events}
\title{set of IMA events.}
\usage{
ima_events()
}
\value{
a vector of strings representing supported IMA events
}
\description{
Returns the set of IMA events.
}
\details{
The set of available events is controlled by available sport plugins. The modules allocated to a user
account control the plugins that are available. Modules are allocated by Catapult support staff.
See \emph{OpenField Cloud API} vignette for more details.
}
\examples{
all_events <- ima_events()
}
