% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetActivities}
\alias{ofCloudGetActivities}
\alias{ofCloudGetActivitiesEx}
\alias{ofCloudGetPeriods}
\alias{ofCloudGetActivity}
\title{get activities}
\usage{
ofCloudGetActivities(credentials, from = NA, to = NA, syncStatus = FALSE)

ofCloudGetActivitiesEx(credentials, from, to, syncStatus = FALSE)

ofCloudGetPeriods(credentials, activity_id)

ofCloudGetActivity(credentials, activity_id, syncStatus = FALSE)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{from, to}{POSIX time in seconds since the start of the epoch. \code{NA} by default.
Can be specified either as \code{(from, to)} or as a half-open interval \code{(from,)} or \code{(,to)}.}

\item{syncStatus}{if \code{TRUE} then the returned data frame has two additional columns, \code{synced} and \code{unsynced}, with the number of SD files in activity that are respectively synchronised / not synchronised.}
}
\value{
\code{ofCloudGetActivities} returns a data frame with activities for a given time interval, or \code{NULL} if there are no activities.

\code{ofCloudGetActivitiesEx} returns a list with an entry for each \code{(from[j], to[j])} interval, 
with the entry content matching ofCloudGetActivities() return value.
The function does not throw if a HTTP call fails for any interval. \code{error} will be present in the returned list in an entry 
for an interval with a failed call.

\code{ofCloudGetPeriods} returns a data frame with period details for a given activity, or \code{NULL} if there are no periods.

\code{ofCloudGetActivity} returns a list with details for a given activity.
}
\description{
\code{ofCloudGetActivities} returns a data frame with all activities associated with the OF account, optionally filtered by a time range (\code{from, to}).
}
\details{
The returned data frame includes the following attributes:\cr
  \code{id, game_id, name, start_time, end_time, periods, tags, tag_list}.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetActivitiesEx()}: get activities. The function treats \code{from}, \code{to} as vectors and triggers one cloud API call 
for each \code{(from[j], to[j])} interval in parallel. 

Be mindful not to overload the server with excessive number of requests. Keep the number of intervals passed into
\code{ofCloudGetActivitiesEx} reasonable. Otherwise, the requests might be throttled on the server.

\item \code{ofCloudGetPeriods()}: get periods for an activity

\item \code{ofCloudGetActivity()}: get details for an activity

}}
\examples{
\dontrun{
to <- as.integer(Sys.time()); from <- to - 7*24*60*60
activities <- ofCloudGetActivities(credentials, from, to)
}
}
