% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetActivityEfforts}
\alias{ofCloudGetActivityEfforts}
\alias{ofCloudGetPeriodEfforts}
\title{get specified efforts for the specified athlete and activity}
\usage{
ofCloudGetActivityEfforts(
  credentials,
  athlete_id,
  activity_id,
  velocityOrAcceleration,
  bands = NA,
  stream_type = NA,
  from = NA,
  to = NA
)

ofCloudGetPeriodEfforts(
  credentials,
  athlete_id,
  period_id,
  velocityOrAcceleration,
  bands = NA,
  stream_type = NA,
  from = NA,
  to = NA
)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{athlete_id}{athlete id as returned by \code{\link{ofCloudGetAthletes}}, \code{\link{ofCloudGetAthletesInActivity}}, \code{\link{ofCloudGetAthletesInPeriod}}, 
\code{\link{ofCloudGetAthleteDevicesInActivity}} or \code{\link{ofCloudGetAthleteDevicesInPeriod}}}

\item{activity_id}{activity id as returned by \code{\link{ofCloudGetActivities}}}

\item{velocityOrAcceleration}{\code{TRUE} to request gen. 2 velocity efforts, \code{FALSE} to request gen. 2 acceleration efforts.}

\item{bands}{vector of required bands. Valid bands are: 1 to 8 for velocities; -3, -2, -1 for decelaration; 1, 2, 3 for acceleration.}

\item{stream_type}{the required stream type; the valid types are \code{"gps"} and \code{"lps"}. By default, a stream type mapped to the specified activity or period will be used.}

\item{from, to}{POSIX time in seconds since the start of the epoch, \code{NA} by default,
can be specified either as \code{(from, to)} or as a half-open interval \code{(from,)} / \code{(,to)}.
If \code{from} is not \code{NA}, then only efforts with the \code{start time >= from} are returned.
If \code{to} is not \code{NA}, then only efforts with the \code{end_time <= to} are returned.}

\item{period_id}{period ID as returned by \code{\link{ofCloudGetPeriods}} or \code{\link{ofCloudGetActivities}}}
}
\value{
ofCloudGetActivityEfforts() returns the specified efforts, or \code{NULL} if there are no data to return.
}
\description{
\code{ofCloudGetActivityEfforts} returns efforts for a given athlete and a given activity. A relevant module (\code{Gen2VelocityBands.Released} or \code{Gen2AccelerationBands.Released}) should be allocated to your account by Catapult support, otherwise respective efforts are not available. Check modules with {\link{ofCloudGetModules}}. Once a module is allocated, reprocess your data through console and complete a full sync to upload efforts into the cloud.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetPeriodEfforts()}: get specified efforts for the specified athlete and period

}}
\seealso{
Other cloud data access APIs: 
\code{\link{ofCloudGetActivityEvents}()},
\code{\link{ofCloudGetActivitySensorData}()}
}
\concept{cloud data access APIs}
