% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetActivityEvents}
\alias{ofCloudGetActivityEvents}
\alias{ofCloudGetPeriodEvents}
\title{get specified events for the specified athlete and activity}
\usage{
ofCloudGetActivityEvents(
  credentials,
  athlete_id,
  activity_id,
  events,
  from = NA,
  to = NA,
  ima_acceleration_intensity_threshold = 0
)

ofCloudGetPeriodEvents(
  credentials,
  athlete_id,
  period_id,
  events,
  from = NA,
  to = NA,
  ima_acceleration_intensity_threshold = 0
)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{athlete_id}{athlete id as returned by \code{\link{ofCloudGetAthletes}}, \code{\link{ofCloudGetAthletesInActivity}},
\code{\link{ofCloudGetAthletesInPeriod}}, \code{\link{ofCloudGetAthleteDevicesInActivity}} or \code{\link{ofCloudGetAthleteDevicesInPeriod}}}

\item{activity_id}{activity id as returned by \code{\link{ofCloudGetActivities}}}

\item{events}{vector of event types, valid vector entries are in \code{\link{ima_events}}; could be specified as numeric IDs (for advanced users)}

\item{from, to}{POSIX time in seconds since the start of the epoch, \code{NA} by default,
can be specified either as \code{(from, to)} or as a half-open interval \code{(from,)} / \code{(,to)}.
If \code{from} is not \code{NA}, then only events with the \code{start time >= from} are returned.
If \code{to} is not \code{NA}, then only events with the \code{end_time <= to} are returned.}

\item{ima_acceleration_intensity_threshold}{intensity threshold for \code{'ima_acceleration'} events, default \code{0}}

\item{period_id}{period ID as returned by \code{\link{ofCloudGetPeriods}} or \code{\link{ofCloudGetActivities}}}
}
\value{
\code{ofCloudGetActivityEvents} returns the specified events (as a named list of dataframes, one entry per one event type), or \code{NULL} if there are no data to return.
}
\description{
\code{ofCloudGetActivityEvents} returns events for a given athlete and a given activity.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetPeriodEvents()}: get specified events for the specified athlete and period

}}
\seealso{
Other cloud data access APIs: 
\code{\link{ofCloudGetActivityEfforts}()},
\code{\link{ofCloudGetActivitySensorData}()}
}
\concept{cloud data access APIs}
