% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetAthletes}
\alias{ofCloudGetAthletes}
\alias{ofCloudGetAthletesInActivity}
\alias{ofCloudGetAthletesInPeriod}
\title{get all athletes associated with the OF account}
\usage{
ofCloudGetAthletes(credentials)

ofCloudGetAthletesInActivity(credentials, activity_id)

ofCloudGetAthletesInPeriod(credentials, period_id)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{activity_id}{activity ID as returned by \code{\link{ofCloudGetActivities}}}

\item{period_id}{period ID as returned by \code{\link{ofCloudGetPeriods}} or \code{\link{ofCloudGetActivities}}}
}
\value{
\code{ofCloudGetAthletes} returns all athletes associated with the OF account, or \code{NULL} if there are no athletes.

\code{ofCloudGetAthletesInActivity} and \code{ofCloudGetAthletesInPeriod} return a data frame with athletes in the specified activity or period.
The functions throw if the HTTP response indicates failure. In particular, it throws \strong{404} if athletes are not mapped to the specified activity or period, see an example in \code{\link{ofCloudParseError}}.
}
\description{
\code{ofCloudGetAthletes} returns all athletes associated with the OF account
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetAthletesInActivity()}: get all athletes for a specified activity.\cr
\code{\link{ofCloudGetAthleteDevicesInActivity}} returns similar data, and additionally device id.

\item \code{ofCloudGetAthletesInPeriod()}: get all athletes for a specified period.\cr
\code{\link{ofCloudGetAthleteDevicesInPeriod}} returns similar data, and additionally device id.

}}
\examples{
\dontrun{
df <- ofCloudGetAthletes(credentials)
}
}
\seealso{
Other athlete details access APIs: 
\code{\link{ofCloudGetAthleteDevicesInActivity}()}
}
\concept{athlete details access APIs}
