% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetBands}
\alias{ofCloudGetBands}
\alias{ofCloudGetBand}
\title{get bands for a given athlete}
\usage{
ofCloudGetBands(credentials, athlete_id)

ofCloudGetBand(credentials, athlete_id, sBand)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{athlete_id}{athlete id as returned by \code{\link{ofCloudGetAthletes}}, \code{\link{ofCloudGetAthletesInActivity}}, \code{\link{ofCloudGetAthletesInPeriod}}, 
\code{\link{ofCloudGetAthleteDevicesInActivity}} or \code{\link{ofCloudGetAthleteDevicesInPeriod}}.}

\item{sBand}{a band name as available from the data frame returned by \code{\link{ofCloudGetBands}}, \emph{e.g.} \code{'Gen2Acceleration'}, \code{'Velocity2'}}
}
\value{
\code{ofCloudGetBands} returns a data frame with bands.

\code{ofCloudGetBands} returns a data frame with band details.\cr
The rows with \code{end_time} attribute equal \code{0} correspond to the current (as opposite to historical) band.
}
\description{
ofCloudGetBands() returns a data frame with bands.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetBand}: get band details for a given athlete and a band
}}

\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetParameters}()},
\code{\link{ofCloudGetStatistics}()},
\code{\link{ofCloudGetTeams}()},
\code{\link{ofCloudGetUserSettings}()},
\code{\link{ofCloudGetVenues}()}
}
\concept{cloud metadata access APIs}
