% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetParameters}
\alias{ofCloudGetParameters}
\title{get all parameters associated with the OF account}
\usage{
ofCloudGetParameters(credentials)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}
}
\value{
\code{ofCloudGetParameters} returns all parameters associated with the OF account, or \code{NULL} if there are no parameters (this should not happen).
}
\description{
Returns all parameters associated with the OF account.
}
\examples{
\dontrun{
df <- ofCloudGetParameters(credentials)
}
}
\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetBands}()},
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetStatistics}()},
\code{\link{ofCloudGetTeams}()},
\code{\link{ofCloudGetUserSettings}()},
\code{\link{ofCloudGetVenues}()}
}
\concept{cloud metadata access APIs}
