% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetStatistics}
\alias{ofCloudGetStatistics}
\title{return the specified statistics}
\usage{
ofCloudGetStatistics(
  credentials,
  params,
  groupby = "athlete",
  filters,
  annotationStats = FALSE
)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{params}{parameters to return, like \code{'total_player_load'}, \code{'total_duration'}, \code{'total_distance'} etc.\cr
All available parameters are returned by \code{\link{ofCloudGetParameters}}.}

\item{groupby}{how to group the results, an element from a set \code{'athlete', 'activity', 'period', 'position', 'team'},\cr
or any combination of the elements. \code{'athlete'} by default. If \code{annotationStats == TRUE} then \code{'annotation_category', 'annotation'} are also valid \code{groupby} values.}

\item{filters}{a list with one or more filters. Each filter is a list with three elements,\cr
1. \code{name} - a field to filter, from the set \code{'date', 'period_id', 'activity_id', 'athlete_id', 'position_id', 'team_id', 'athlete_group_id', 'lastActivities', 'timeRange', 'period_name', 'snapshot_name', 'rotationNumber', 'tag_id', 'month_id', 'day_id'},\cr
2. \code{comparison} - what type of filter to use \code{(=, <>, >, <, >=, <=)},\cr
3. \code{values} - values for the filter, \emph{e.g.} an activity ID(s) for the \code{'activity_id'} filter.}

\item{annotationStats}{query annotation database, default \code{FALSE}.\cr
To group by annotation_category and/or annotation (annotation name) annotationStats must be \code{TRUE}.}
}
\value{
\code{ofCloudGetStatistics} returns a data frame with the specified statistics, or \code{NULL} if there are no rows to return.
}
\description{
Returns a data frame with the specified statistics. Use \code{\link{ofCloudGetUserSettings}} to retrieve velocity and distance units.
}
\examples{
\dontrun{
activities <- ofCloudGetActivities(token)
grouping <- c("athlete", "activity", "period")
parameters <- c("total_player_load", "total_distance")
filter1 <- list("name"="activity_id", "comparison" = "=", "values" = c(id1, id2))
filter2 <- list("name"="athlete_name", "comparison" = "=", "values" = "Athlete Name")
df1 <- ofCloudGetStatistics(credentials = token, params = parameters, groupby = grouping, filters = filter1) 
df2 <- ofCloudGetStatistics(credentials = token, params = parameters, groupby = grouping, filters = filter2) 
df3 <- ofCloudGetStatistics(credentials = token, params = parameters, groupby = grouping, filters = c(filter1, filter2)) 
}
}
\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetBands}()},
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetParameters}()},
\code{\link{ofCloudGetTeams}()},
\code{\link{ofCloudGetUserSettings}()},
\code{\link{ofCloudGetVenues}()}
}
\concept{cloud metadata access APIs}
