% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetTeams}
\alias{ofCloudGetTeams}
\title{get teams on the account}
\usage{
ofCloudGetTeams(credentials)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}
}
\value{
\code{ofCloudGetTeams} returns a data frame with teams.
}
\description{
Returns a data frame with teams.
}
\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetBands}()},
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetParameters}()},
\code{\link{ofCloudGetStatistics}()},
\code{\link{ofCloudGetUserSettings}()},
\code{\link{ofCloudGetVenues}()}
}
\concept{cloud metadata access APIs}
