% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetToken}
\alias{ofCloudGetToken}
\alias{safe_ofCloudGetToken}
\alias{safe_ofCloudGetTokenEx}
\alias{safe_ofCloudValidateToken}
\alias{ofCloudCreateToken}
\title{OF cloud login}
\usage{
ofCloudGetToken(
  sRegion,
  sName,
  sPwd,
  sStage = "main",
  sClientID,
  sClientSecret,
  traceURL = FALSE
)

safe_ofCloudGetToken(
  sRegion,
  sName,
  sPwd,
  sStage = "main",
  sClientID,
  sClientSecret,
  traceURL = FALSE
)

safe_ofCloudGetTokenEx(
  sName,
  sPwd,
  sStage = "main",
  sClientID,
  sClientSecret,
  traceURL = FALSE
)

safe_ofCloudValidateToken(credentials, minTimeToExpiry = 3)

ofCloudCreateToken(
  sToken,
  sRegion,
  sStage = "main",
  tokenExpireTime = 0,
  traceURL = FALSE,
  refresh_token = NULL
)
}
\arguments{
\item{sRegion}{a region from the set \code{"APAC", "EMEA", "America", "China"}}

\item{sName}{a User Name}

\item{sPwd}{a User Password}

\item{sStage}{a production (\code{"main"}, default), or a test stage (\code{"alpha", "beta"})}

\item{sClientID, sClientSecret}{should keep default values unless specifically instructed by Catapult Sports}

\item{traceURL}{controls API tracing}

\item{credentials}{R6 object as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{minTimeToExpiry}{minimum time required before the token expires (seconds)}

\item{sToken}{a token string}

\item{tokenExpireTime}{token expire time as POSIX time in seconds since the start of the epoch, designed to match the token Expiration Date from the OF Cloud -> API Tokens screen}

\item{refresh_token}{a refresh token string}
}
\value{
\code{ofCloudGetToken} returns ofCredentials R6 object.

\code{safe_ofCloudGetToken} returns a list with \code{result} with ofCredentials R6 object if successful 
or \code{error} if failed.

\code{safe_ofCloudGetTokenEx} returns a list with \code{result} with ofCredentials R6 object if successful 
or \code{error} if failed.

\code{safe_ofCloudValidateToken} returns a list with \code{result} with ofCredentials R6 object if successful 
or \code{error} if failed.

\code{ofCloudCreateToken} returns ofCredentials R6 object.
}
\description{
Implements OF cloud login with basic authentication, without appending user name and password to URL.
}
\section{Functions}{
\itemize{
\item \code{safe_ofCloudGetToken}: a safe version of the function that returns a list with 'error' and 'result'

\item \code{safe_ofCloudGetTokenEx}: OF cloud login when the region is not known, by trying all possible regions

\item \code{safe_ofCloudValidateToken}: refresh token if necessary

Validates a token in case it has expired.

\item \code{ofCloudCreateToken}: create ofCredentials object from a given token string

The token object created with \code{ofCloudCreateToken} cannot be renewed with \code{safe_ofCloudValidateToken} unless \code{refresh_token} is provided.
}}

\examples{
token <- ofCloudGetToken("APAC", "sergey", "password", "clientID", "clientSecret")
\dontrun{
 safe_ofCloudValidateToken(credentials)
}
}
