% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetUserSettings}
\alias{ofCloudGetUserSettings}
\title{get user settings}
\usage{
ofCloudGetUserSettings(credentials)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}
}
\value{
\code{ofCloudGetUserSettings} returns a data frame with user settings.
}
\description{
\code{ofCloudGetUserSettings} returns a data frame with user settings; the settings control the units of velocity and distance as retrieved by \code{\link{ofCloudGetStatistics}}.
}
\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetBands}()},
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetParameters}()},
\code{\link{ofCloudGetStatistics}()},
\code{\link{ofCloudGetTeams}()},
\code{\link{ofCloudGetVenues}()}
}
\concept{cloud metadata access APIs}
