% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetVenues}
\alias{ofCloudGetVenues}
\alias{ofCloudGetVenue}
\title{get venues}
\usage{
ofCloudGetVenues(credentials)

ofCloudGetVenue(credentials, id)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{id}{a venue id as available from \code{ofCloudGetVenues}}
}
\value{
\code{ofCloudGetVenues} returns a data frame with venues system attributes.

\code{ofCloudGetVenue} returns a data frame with a given venue system attributes.
}
\description{
\code{ofCloudGetVenues} returns a data frame with venues system attributes.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetVenue}: return a data frame with a given venue system attributes.
}}

\seealso{
Other cloud metadata access APIs: 
\code{\link{ofCloudGetBands}()},
\code{\link{ofCloudGetCustomerInfo}()},
\code{\link{ofCloudGetModules}()},
\code{\link{ofCloudGetParameters}()},
\code{\link{ofCloudGetStatistics}()},
\code{\link{ofCloudGetTeams}()},
\code{\link{ofCloudGetUserSettings}()}
}
\concept{cloud metadata access APIs}
