% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFiles.R
\name{ofDataFileCSV}
\alias{ofDataFileCSV}
\alias{of10HzDataFileCSV}
\alias{ofIMAFileCSV}
\alias{ofVelEffortsFileCSV}
\title{Get CSV data file packaged with catapultR}
\usage{
ofDataFileCSV()

of10HzDataFileCSV()

ofIMAFileCSV()

ofVelEffortsFileCSV()
}
\value{
\code{ofDataFileCSV} returns an example CSV file name with 100 Hz data.

\code{of10HzDataFileCSV} returns an example CSV file name with 10 Hz data.

\code{ofIMAFileCSV} returns an example CSV file name with IMA events.

\code{ofVelEffortsFileCSV} returns an example CSV file name with velocity efforts.
}
\description{
ofDataFileCSV() returns CSV 100 Hz data file name, packaged with catapultR package. 100 Hz data is produced in \emph{OpenField} console by raw files post-processing.
}
\section{Functions}{
\itemize{
\item \code{of10HzDataFileCSV()}: access 10 Hz sample CSV file originated from \code{\link{ofCloudGetActivitySensorData}} / \code{\link{ofCloudGetPeriodSensorData}}.\cr
Facilitates OpenField Cloud API vignette.

\item \code{ofIMAFileCSV()}: access IMA events sample CSV file originated from \code{\link{ofCloudGetActivityEvents}} / \code{\link{ofCloudGetPeriodEvents}}.\cr
Facilitates OpenField Cloud API vignette.

\item \code{ofVelEffortsFileCSV()}: access velocity efforts sample CSV file originated from \code{\link{ofCloudGetActivityEfforts}} / \code{\link{ofCloudGetPeriodEfforts}}.\cr
Facilitates OpenField Cloud API vignette.

}}
\examples{
sCSV <- ofDataFileCSV()
hiFreq <- read_CATcsv(sCSV)
df100Hz <- hiFreq$data
 
}
