% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CATcsv.R
\name{read_CATcsv}
\alias{read_CATcsv}
\title{Read high frequency CSV}
\usage{
read_CATcsv(
  filedatapath,
  col_types = paste0(rep("d", 26), collapse = ""),
  suppressMsgs = TRUE
)
}
\arguments{
\item{filedatapath}{source CSV file.}

\item{col_types}{column specification for \code{readr::read_csv}, keeping the default value is recommended.}

\item{suppressMsgs}{\code{TRUE} to suppress messages from \code{readr::read_csv}.}
}
\value{
a list with metadata and \code{data} tibble.
}
\description{
Loads high frequency parameters for analytics CSV as a data_tibble.\cr
\code{read_CATcsv} supplements \code{\link{write_CATcsv}}.
}
\examples{
x <- read_CATcsv(ofDataFileCSV())
xData <- x$data
posixTimeStamp <- x$CentisecTime / 100   # time in seconds since the start of the epoch
}
\seealso{
Other read and write Catapult 100 Hz CSV files: 
\code{\link{write_CATcsv}()}
}
\concept{read and write Catapult 100 Hz CSV files}
