% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CATcsv.R
\name{write_CATcsv}
\alias{write_CATcsv}
\title{Write high frequency CSV, preserving the metadata}
\usage{
write_CATcsv(x, targetCSV)
}
\arguments{
\item{x}{a list with metadata and the 100 Hz data, as returned by \code{\link{read_CATcsv}}.}

\item{targetCSV}{target CSV file.}
}
\value{
\code{x}, invisibly.
}
\description{
Writes high frequency data into a CSV file, preserving the metadata.\cr
Use this function to save 100 Hz data including the metadata when the data was updated.\cr
\code{write_CATcsv} supplements \code{\link{read_CATcsv}}.
}
\seealso{
Other read and write Catapult 100 Hz CSV files: 
\code{\link{read_CATcsv}()}
}
\concept{read and write Catapult 100 Hz CSV files}
