% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofCloud.R
\name{ofCloudGetAthleteDevicesInActivity}
\alias{ofCloudGetAthleteDevicesInActivity}
\alias{ofCloudGetAthleteDevicesInPeriod}
\title{get device id, athlete id and other details for a given activity or period}
\usage{
ofCloudGetAthleteDevicesInActivity(credentials, activity_id)

ofCloudGetAthleteDevicesInPeriod(credentials, period_id)
}
\arguments{
\item{credentials}{as returned by \code{\link{ofCloudGetToken}}, \code{\link{safe_ofCloudGetToken}} or \code{\link{safe_ofCloudGetTokenEx}}}

\item{activity_id}{as returned by \code{\link{ofCloudGetActivities}}}

\item{period_id}{as returned by \code{\link{ofCloudGetPeriods}} or \code{\link{ofCloudGetActivities}}}
}
\value{
\code{ofCloudGetAthleteDevicesInActivity} returns a data frame with \code{device_id}, \code{athlete_id} and \code{activity_dictionaries} for a given activity, or \code{NULL} if there are no rows to return.
}
\description{
\code{ofCloudGetAthleteDevicesInActivity} returns a data frame with \code{device_id}, \code{athlete_id} and \code{activity_dictionaries} for a given activity.\cr
\code{\link{ofCloudGetAthletesInActivity}} and \code{\link{ofCloudGetAthletesInPeriod}} return \code{athlete_id} but do not return \code{device_id}.
}
\section{Functions}{
\itemize{
\item \code{ofCloudGetAthleteDevicesInPeriod()}: returns a data frame with device id, athlete and team details for a given period.\cr

}}
\examples{
\dontrun{
to <- as.integer(Sys.time()); from <- to - 7*24*60*60
activities <- ofCloudGetActivities(credentials, from, to)
j <- 1   # the activity of interest
dfAD <- ofCloudGetAthleteDevicesInActivity(credentials, activities$id[j])
dfActDict <- bind_rows(dfAD$activity_dictionaries)
}
}
\seealso{
Other athlete details access APIs: 
\code{\link{ofCloudGetAthletes}()}
}
\concept{athlete details access APIs}
